Extract logic/components from App.cs
