+++ src/RepoCleaner/App.cs
@@ -1,11 +1,7 @@
 ﻿using Develix.RepoCleaner.ConsoleComponents;
-using Develix.RepoCleaner.Git;
-using Develix.RepoCleaner.Git.Model;
 using Develix.RepoCleaner.Model;
-using Develix.RepoCleaner.Store;
 using Develix.RepoCleaner.Store.ConsoleSettingsUseCase;
 using Develix.RepoCleaner.Store.RepositoryInfoUseCase;
-using Develix.RepoCleaner.Utils;
 using Fluxor;
 using Spectre.Console;
 
@@ -38,14 +34,15 @@ public class App
         {
             if (consoleArguments.Config)
             {
-                Config();
+                Config.Show(CredentialName, consoleSettingsState, dispatcher);
                 return;
             }
 
             await InitConsole(consoleArguments, appSettings);
             LogErrors(repositoryInfoState);
             ShowOverviewTable();
-            ShowDeletePrompt();
+            if (consoleSettingsState.Value.ShowDeletePrompt)
+                ShowDeletePrompt();
 
         }
         catch (Exception ex)
@@ -54,60 +51,10 @@ public class App
         }
     }
 
-    private void Config()
-    {
-        var token = AnsiConsole.Prompt(new TextPrompt<string>("Enter [green]azure devops token[/]")
-            .PromptStyle("red")
-            .Secret());
-        AnsiConsole.Status().Start("Storing credentials",
-            async (ctx) =>
-            {
-                var action = new ConfigureCredentialsAction(CredentialName, token);
-                dispatcher.Dispatch(action);
-                await AsyncHelper.WaitUntilAsync(() => !consoleSettingsState.Value.Configuring, 100, 2000, default);
-                ctx.Status("Credentials initialized");
-            });
-    }
-
-    private IReadOnlyList<Branch> GetBranchesToDelete()
-    {
-        var deletableBranches = repositoryInfoState.Value.Repository.Branches.Where(b => IsDeletable(b)).ToList();
-
-        if (deletableBranches.Count == 0)
-        {
-            AnsiConsole.MarkupLine("[grey]No branches can be deleted.[/]");
-            return Array.Empty<Branch>();
-        }
-        return DeleteBranchSelectionPrompt.Show(deletableBranches, repositoryInfoState.Value);
-
-        static bool IsDeletable(Branch b) => !b.IsRemote && !b.IsCurrent;
-    }
-
-    private void Delete(IReadOnlyList<Branch> branchesToDelete)
-    {
-        using var handler = new Handler(repositoryInfoState.Value.Repository);
-        foreach (var branch in branchesToDelete)
-        {
-            var result = handler.TryDeleteBranch(branch);
-            var message = result.Valid
-                ? $"[green]Deleted branch {branch.FriendlyName}[/]"
-                : $"[red]Failed:[/] {result.Message}";
-            AnsiConsole.MarkupLine(message);
-        }
-    }
-
     private async Task InitConsole(ConsoleArguments consoleArguments, AppSettings appSettings)
     {
-        await AnsiConsole.Status()
-            .StartAsync("Loading", async ctx =>
-            {
-                ctx.Spinner(Spinner.Known.Dots);
-                await InitStore(ctx);
-                InitConsoleSettings(consoleArguments, appSettings, ctx);
-                await LoginServices(ctx);
-                await InitRepository(consoleArguments, ctx);
-                ctx.Status = "Done";
-            });
+        var initConsole = new Init(CredentialName, dispatcher, repositoryInfoState, store);
+        await initConsole.Execute(consoleArguments, appSettings);
     }
 
     private static void LogErrors(IState<RepositoryInfoState> repositoryInfoState)
@@ -126,46 +73,7 @@ public class App
 
     private void ShowDeletePrompt()
     {
-        if (consoleSettingsState.Value.ShowDeletePrompt)
-        {
-            var branchesToDelete = GetBranchesToDelete();
-            Delete(branchesToDelete);
-        }
-    }
-
-    private async Task InitStore(StatusContext statusContext)
-    {
-        statusContext.Status = "Initializing";
-        await store.InitializeAsync();
-    }
-
-    private void InitConsoleSettings(ConsoleArguments consoleArguments, AppSettings appSettings, StatusContext statusContext)
-    {
-        statusContext.Status = "Initializing Console Settings";
-        dispatcher.Dispatch(new SetConsoleSettingsAction(consoleArguments, appSettings));
-    }
-
-    private async Task LoginServices(StatusContext statusContext)
-    {
-        statusContext.Status = "Login Services";
-        dispatcher.Dispatch(new LoginServicesAction(CredentialName));
-        await AsyncHelper.WaitUntilAsync(
-            () => IsFinalState(repositoryInfoState.Value.WorkItemServiceState),
-            100,
-            30000,
-            default);
-
-        static bool IsFinalState(ServiceConnectionState state) => state == ServiceConnectionState.Connected || state == ServiceConnectionState.FailedToConnect;
-    }
-
-    private async Task InitRepository(ConsoleArguments consoleArguments, StatusContext statusContext)
-    {
-        statusContext.Status = "Initializing repository and work items";
-        dispatcher.Dispatch(new InitRepositoryAction(consoleArguments.Branches));
-        var waitRepositoryLoaded = AsyncHelper.WaitUntilAsync(() => repositoryInfoState.Value.RepositoryLoaded, 100, 30000, default)
-            .ContinueWith(t => statusContext.Status = "Initializing work items");
-        var waitWorkItemsLoaded = AsyncHelper.WaitUntilAsync(() => repositoryInfoState.Value.WorkItemsLoaded, 100, 30000, default)
-            .ContinueWith(t => statusContext.Status = "Initializing repository");
-        await Task.WhenAll(waitRepositoryLoaded, waitWorkItemsLoaded);
+        var branchesToDelete = Delete.Prompt(repositoryInfoState);
+        Delete.Execute(branchesToDelete, repositoryInfoState);
     }
 }
